🎨 [chore] change API name

+++ src/shared/axios.js
@@ -9,7 +9,6 @@ const verifyJwt = () => {
       throw new Error("There is no jwt");
     }
   } catch (error) {
-    console.log(error);
     return "login";
   }
 
@@ -69,13 +68,32 @@ export const getItemsBySearch = async (path, itemId, keyword) => {
   return result.data;
 };
 
+export const requestLookUp = async () => {
+  const jwt = verifyJwt();
+  if (jwt === "login") {
+    return "login";
+  }
+
+  const result = await axios.get(
+    `${process.env.REACT_APP_BACK_URL}/api/v1/users/look-up`,
+    {
+      headers: { jwt },
+    },
+  );
+  if (verifyLogin(result.data.error) === "login") {
+    return "login";
+  }
+  updateJwt(result.data.response.jwt);
+
+  return result.data.response;
+};
+
 // JWT를 리턴하지 않는 데이터 (비회원 가능)
-// requestPublicData
-export const getInfo = async (path) => {
+export const requestPublicData = async (path) => {
   const jwt = verifyJwt();
 
   const result = await axios.get(
-    `${process.env.REACT_APP_BACK_URL}/api/v1/${path}`,
+    `${process.env.REACT_APP_BACK_URL}/api/${path}`,
     {
       headers: { jwt },
     },
@@ -85,15 +103,14 @@ export const getInfo = async (path) => {
 };
 
 // JWT를 리턴하는 데이터 (회원 전용)
-// requestAuthData
-export const getData = async (path) => {
-  if (verifyJwt() === "login") {
+export const requestAuthData = async (path) => {
+  const jwt = verifyJwt();
+  if (jwt === "login") {
     return "login";
   }
-  const jwt = verifyJwt();
 
   const result = await axios.get(
-    `${process.env.REACT_APP_BACK_URL}/api/v1/${path}`,
+    `${process.env.REACT_APP_BACK_URL}/api/${path}`,
     {
       headers: { jwt },
     },
@@ -109,9 +126,12 @@ export const getData = async (path) => {
 // 회원인지, 비회원인지 체크
 export const checkLoginWithData = async (path) => {
   const jwt = verifyJwt();
+  if (jwt === "login") {
+    return "login";
+  }
 
   const result = await axios.get(
-    `${process.env.REACT_APP_BACK_URL}/api/v1/${path}`,
+    `${process.env.REACT_APP_BACK_URL}/api/${path}`,
     {
       headers: { jwt },
     },
@@ -127,9 +147,12 @@ export const checkLoginWithData = async (path) => {
 // get 요청
 export const getAction = async (path) => {
   const jwt = verifyJwt();
+  if (jwt === "login") {
+    return "login";
+  }
 
   const result = await axios.get(
-    `${process.env.REACT_APP_BACK_URL}/api/v1/${path}`,
+    `${process.env.REACT_APP_BACK_URL}/api/${path}`,
     {
       headers: { jwt },
     },
@@ -145,9 +168,12 @@ export const getAction = async (path) => {
 // delete 요청
 export const deleteAction = async (path) => {
   const jwt = verifyJwt();
+  if (jwt === "login") {
+    return "login";
+  }
 
   const result = await axios.delete(
-    `${process.env.REACT_APP_BACK_URL}/api/v1/${path}`,
+    `${process.env.REACT_APP_BACK_URL}/api/${path}`,
     {
       headers: { jwt },
     },
@@ -163,9 +189,12 @@ export const deleteAction = async (path) => {
 // post 요청
 export const postAction = async (path, json) => {
   const jwt = verifyJwt();
+  if (jwt === "login") {
+    return "login";
+  }
 
   const result = await axios.post(
-    `${process.env.REACT_APP_BACK_URL}/api/v1/${path}`,
+    `${process.env.REACT_APP_BACK_URL}/api/${path}`,
     JSON.stringify(json),
     {
       headers: { jwt, "Content-Type": "application/json" },
@@ -182,9 +211,12 @@ export const postAction = async (path, json) => {
 // post(image) 요청
 export const postImgAction = async (path, file) => {
   const jwt = verifyJwt();
+  if (jwt === "login") {
+    return "login";
+  }
 
   const result = await axios.post(
-    `${process.env.REACT_APP_BACK_URL}/api/v1/${path}`,
+    `${process.env.REACT_APP_BACK_URL}/api/${path}`,
     file,
     {
       headers: { jwt },
@@ -201,8 +233,11 @@ export const postImgAction = async (path, file) => {
 // put 요청
 export const putAction = async (path, data) => {
   const jwt = verifyJwt();
+  if (jwt === "login") {
+    return "login";
+  }
 
-  const url = `${process.env.REACT_APP_BACK_URL}/api/v1/${path}`;
+  const url = `${process.env.REACT_APP_BACK_URL}/api/${path}`;
   const options = { headers: { jwt } };
   const result = await axios.put(url, data, options);
 
@@ -217,9 +252,12 @@ export const putAction = async (path, data) => {
 // patch 요청
 export const patchAction = async (path, json) => {
   const jwt = verifyJwt();
+  if (jwt === "login") {
+    return "login";
+  }
 
   const result = await axios.patch(
-    `${process.env.REACT_APP_BACK_URL}/api/v1/${path}`,
+    `${process.env.REACT_APP_BACK_URL}/api/${path}`,
     {
       headers: { jwt },
     },
@@ -235,9 +273,12 @@ export const patchAction = async (path, json) => {
 // patch JSON 요청
 export const patchJsonAction = async (path, json) => {
   const jwt = verifyJwt();
+  if (jwt === "login") {
+    return "login";
+  }
 
   const result = await axios.patch(
-    `${process.env.REACT_APP_BACK_URL}/api/v1/${path}`,
+    `${process.env.REACT_APP_BACK_URL}/api/${path}`,
     JSON.stringify(json),
     {
       headers: { jwt, "Content-Type": "application/json" },
